split license / fileInfo routes and tests
