+++ routes/licenses.js
@@ -1,6 +1,5 @@
 const Joi = require('joi');
 
-const errors = require('../services/util/errors');
 const { license: serialize } = require('../services/util/serializers');
 
 const routes = [];
@@ -15,19 +14,6 @@ const licenseSchema = Joi.object({
     .required()
     .items(Joi.string()),
 });
-
-function handleError(h, { message }) {
-  switch (message) {
-    case errors.invalidUrl:
-      return h.error(message, { statusCode: 422 });
-    case errors.emptyResponse:
-      return h.error(message, { statusCode: 404 });
-    case errors.apiUnavailabe:
-      return h.error(message, { statusCode: 503 });
-    default:
-      return h.error(message);
-  }
-}
 
 routes.push({
   path: '/licenses/compatible/{licenseId}',
@@ -69,35 +55,6 @@ routes.push({
     const licenses = licenseStore.all();
     const response = licenses.map(serialize);
     return h.response(response);
-  },
-});
-
-routes.push({
-  path: '/license/{fileUrlOrTitle}',
-  method: 'GET',
-  options: {
-    description: 'Image license',
-    notes: 'Returns the most liberal license for the given image',
-    validate: {
-      params: {
-        fileUrlOrTitle: Joi.string(),
-      },
-    },
-    response: {
-      schema: licenseSchema,
-    },
-  },
-  handler: async (request, h) => {
-    const { fileData, licenses } = request.server.app.services;
-    const { fileUrlOrTitle } = request.params;
-    try {
-      const { title, wikiUrl } = await fileData.getFileData(fileUrlOrTitle);
-      const license = await licenses.getLicense({ title, wikiUrl });
-      const response = serialize(license);
-      return h.response(response);
-    } catch (error) {
-      return handleError(h, error);
-    }
   },
 });
